Services are confusing.
I've not yet achieved making this a functional service. It's a
step that continues to alude me. I think I'm going to have to
devote a decent amount of time to the study of how linux services
are made and then I might be able to get this working the way I
want it to.

The good thing is that the functional side of it works just fine.