Maybe a change in direction is needed.
I'm saving what I've got here now, but I think I'm going to take
this project in a bit of a different direction. I'm not going to
bother with turning this into a service, but rather I will try to
make it easy to use so that if a service is what you want, you can
make it your self. I'll just make this a library that is easy to
work with. And maybe a docker container.