Progress towards a more sensible library.
I've started to break things up in a sensible way to make this a
library that can be reused more and hopefully distributed to the
world at large. It's got some ways to go yet, but right now I need
to do some sys admin stuff with my repos so I need this code
commited so I can move it.