Just needs a touch of polish.
It's time for bed. But I got this basically to where I wanted it.
It is generalized so you can now setup a system relatively easily.
There is a bit of effort getting the interface setup properly, but
assuming you can do that, the rest is easy.

I expanded and updated the README.md preparing for a github
release in the near future.